refactor to move wait_for_deployment to a common utils.py
